+++ src/Contracts/ExceptionHandlerInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Codemonster\Annabel\Contracts;
+
+use Throwable;
+use Codemonster\Http\Response;
+
+interface ExceptionHandlerInterface
+{
+    public function handle(Throwable $e): Response;
+}